+++ modules/analysis/aws-batch-demo/images/processing-mock/src/processing_mock/processing.py
@@ -35,7 +35,9 @@ def main(table_name, index, batch_id, s3bucketout) -> int:
     ).get("Item", {})
 
     if not item:
-        raise Exception(f"pk: {batch_id} sk: {index} not existing in table: {table_name}")
+        raise Exception(
+            f"pk: {batch_id} sk: {index} not existing in table: {table_name}"
+        )
 
     # DO SOME PROCESSING HERE
     s3 = boto3.client("s3")