+++ modules/post-processing/yolo-object-detection/stack.py
@@ -13,17 +13,18 @@
 #    limitations under the License.
 
 import logging
-from typing import Any, Dict, cast
 import os
+from typing import Any, Dict, cast
+
+import aws_cdk.aws_batch_alpha as batch
 import aws_cdk.aws_ec2 as ec2
 import aws_cdk.aws_ecr as ecr
 import aws_cdk.aws_ecs as ecs
 import aws_cdk.aws_iam as iam
-import aws_cdk.aws_batch_alpha as batch
 from aws_cdk import Duration, Stack, Tags
-from constructs import Construct, IConstruct
 from aws_cdk.aws_ecr_assets import DockerImageAsset
-from cdk_ecr_deployment import ECRDeployment, DockerImageName
+from cdk_ecr_deployment import DockerImageName, ECRDeployment
+from constructs import Construct, IConstruct
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
@@ -53,19 +54,25 @@ class ObjectDetection(Stack):  # type: ignore
         dep_mod = f"addf-{deployment_name}-{module_name}"
 
         self.repository_name = dep_mod
-        repo = ecr.Repository(self, id=self.repository_name, repository_name=self.repository_name)
+        repo = ecr.Repository(
+            self, id=self.repository_name, repository_name=self.repository_name
+        )
         self.image_uri = f"{repo.repository_uri}:latest"
 
         policy_statements = [
             iam.PolicyStatement(
                 actions=["dynamodb:*"],
                 effect=iam.Effect.ALLOW,
-                resources=[f"arn:aws:dynamodb:{self.region}:{self.account}:table/addf*"],
+                resources=[
+                    f"arn:aws:dynamodb:{self.region}:{self.account}:table/addf*"
+                ],
             ),
             iam.PolicyStatement(
                 actions=["ecr:*"],
                 effect=iam.Effect.ALLOW,
-                resources=[f"arn:aws:ecr:{self.region}:{self.account}:repository/{dep_mod}*"],
+                resources=[
+                    f"arn:aws:ecr:{self.region}:{self.account}:repository/{dep_mod}*"
+                ],
             ),
             iam.PolicyStatement(
                 actions=["s3:GetObject", "s3:GetObjectAcl", "s3:ListBucket"],
@@ -83,11 +90,15 @@ class ObjectDetection(Stack):  # type: ignore
             ),
             inline_policies={"DagPolicyDocument": dag_document},
             managed_policies=[
-                iam.ManagedPolicy.from_aws_managed_policy_name("service-role/AmazonECSTaskExecutionRolePolicy"),
+                iam.ManagedPolicy.from_aws_managed_policy_name(
+                    "service-role/AmazonECSTaskExecutionRolePolicy"
+                ),
                 iam.ManagedPolicy.from_managed_policy_arn(
                     self, id="fullaccess", managed_policy_arn=s3_access_policy
                 ),
-                iam.ManagedPolicy.from_aws_managed_policy_name("AmazonSageMakerFullAccess"),
+                iam.ManagedPolicy.from_aws_managed_policy_name(
+                    "AmazonSageMakerFullAccess"
+                ),
             ],
             max_session_duration=Duration.hours(12),
         )