+++ modules/sensor-extraction/ros-to-parquet/stack.py
@@ -13,17 +13,18 @@
 #    limitations under the License.
 
 import logging
-from typing import Any, Dict, cast
 import os
+from typing import Any, Dict, cast
+
+import aws_cdk.aws_batch_alpha as batch
 import aws_cdk.aws_ec2 as ec2
 import aws_cdk.aws_ecr as ecr
 import aws_cdk.aws_ecs as ecs
 import aws_cdk.aws_iam as iam
-import aws_cdk.aws_batch_alpha as batch
 from aws_cdk import Duration, Stack, Tags
-from constructs import Construct, IConstruct
 from aws_cdk.aws_ecr_assets import DockerImageAsset
-from cdk_ecr_deployment import ECRDeployment, DockerImageName
+from cdk_ecr_deployment import DockerImageName, ECRDeployment
+from constructs import Construct, IConstruct
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
@@ -50,7 +51,7 @@ class RosToParquetBatchJob(Stack):  # type: ignore
             **kwargs,
         )
 
-        if platform not in ['FARGATE', 'EC2']:
+        if platform not in ["FARGATE", "EC2"]:
             raise ValueError
 
         Tags.of(scope=cast(IConstruct, self)).add(
@@ -61,15 +62,14 @@ class RosToParquetBatchJob(Stack):  # type: ignore
         dep_mod = f"addf-{deployment_name}-{module_name}"
 
         self.repository_name = dep_mod
-        repo = ecr.Repository(self, id=self.repository_name, repository_name=self.repository_name)
+        repo = ecr.Repository(
+            self, id=self.repository_name, repository_name=self.repository_name
+        )
 
         local_image = DockerImageAsset(
             self,
             "RosToParquet",
-            directory=os.path.join(
-                os.path.dirname(os.path.abspath(__file__)),
-                "src"
-            ),
+            directory=os.path.join(os.path.dirname(os.path.abspath(__file__)), "src"),
         )
 
         image_uri = f"{repo.repository_uri}:latest"
@@ -84,12 +84,16 @@ class RosToParquetBatchJob(Stack):  # type: ignore
             iam.PolicyStatement(
                 actions=["dynamodb:*"],
                 effect=iam.Effect.ALLOW,
-                resources=[f"arn:aws:dynamodb:{self.region}:{self.account}:table/addf*"],
+                resources=[
+                    f"arn:aws:dynamodb:{self.region}:{self.account}:table/addf*"
+                ],
             ),
             iam.PolicyStatement(
                 actions=["ecr:*"],
                 effect=iam.Effect.ALLOW,
-                resources=[f"arn:aws:ecr:{self.region}:{self.account}:repository/{dep_mod}*"],
+                resources=[
+                    f"arn:aws:ecr:{self.region}:{self.account}:repository/{dep_mod}*"
+                ],
             ),
             iam.PolicyStatement(
                 actions=["s3:GetObject", "s3:GetObjectAcl", "s3:ListBucket"],
@@ -107,7 +111,9 @@ class RosToParquetBatchJob(Stack):  # type: ignore
             ),
             inline_policies={"DagPolicyDocument": dag_document},
             managed_policies=[
-                iam.ManagedPolicy.from_aws_managed_policy_name("service-role/AmazonECSTaskExecutionRolePolicy"),
+                iam.ManagedPolicy.from_aws_managed_policy_name(
+                    "service-role/AmazonECSTaskExecutionRolePolicy"
+                ),
                 iam.ManagedPolicy.from_managed_policy_arn(
                     self, id="fullaccess", managed_policy_arn=s3_access_policy
                 ),
@@ -122,18 +128,21 @@ class RosToParquetBatchJob(Stack):  # type: ignore
                 image=ecs.ContainerImage.from_ecr_repository(repo, "latest"),
                 command=["bash", "entrypoint.sh"],
                 environment={
-                        "AWS_DEFAULT_REGION": self.region,
-                        "AWS_ACCOUNT_ID": self.account,
-                        "DEBUG": "true",
+                    "AWS_DEFAULT_REGION": self.region,
+                    "AWS_ACCOUNT_ID": self.account,
+                    "DEBUG": "true",
                 },
                 job_role=role,
                 execution_role=role,
                 memory_limit_mib=memory_limit_mib,
-                vcpus=vcpus
+                vcpus=vcpus,
             ),
             job_definition_name=self.repository_name,
-            platform_capabilities=[batch.PlatformCapabilities.FARGATE \
-                if platform == "FARGATE" else batch.PlatformCapabilities.EC2],
+            platform_capabilities=[
+                batch.PlatformCapabilities.FARGATE
+                if platform == "FARGATE"
+                else batch.PlatformCapabilities.EC2
+            ],
             retry_attempts=retries,
             timeout=Duration.seconds(timeout_seconds),
         )